33% coverage

+++ rest-service/src/test/java/org/veupathdb/service/multiblast/model/blast/BlastNConfigTest.java
@@ -152,7 +152,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setMaxHSPs((short) 42);
+      target.setMaxHSPs(42);
       target.toArgs(build);
       assertEquals("-max_hsps='42'", build.toString());
     }
@@ -174,7 +174,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setNumDescriptions((short) 42);
+      target.setNumDescriptions(42);
       target.toArgs(build);
       assertEquals("-num_descriptions='42'", build.toString());
     }
@@ -306,7 +306,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setWindowSize((short) 450);
+      target.setWindowSize(450);
       target.toArgs(build);
       assertEquals("-window_size='450'", build.toString());
     }
@@ -475,7 +475,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setWordSize((short) 420);
+      target.setWordSize(420);
       target.toArgs(build);
       assertEquals("-word_size='420'", build.toString());
     }
@@ -530,7 +530,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setGapOpen((byte) 33);
+      target.setGapOpen(33);
       target.toArgs(build);
       assertEquals("-gapopen='33'", build.toString());
     }
@@ -541,7 +541,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setGapExtend((byte) 77);
+      target.setGapExtend(77);
       target.toArgs(build);
       assertEquals("-gapextend='77'", build.toString());
     }
@@ -552,7 +552,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setPenalty((byte) 77);
+      target.setPenalty(77);
       target.toArgs(build);
       assertEquals("-penalty='77'", build.toString());
     }
@@ -563,7 +563,7 @@ class BlastNConfigTest
       var target = new BlastNConfig();
       var build  = new CliBuilder();
 
-      target.setReward((byte) 77);
+      target.setReward(77);
       target.toArgs(build);
       assertEquals("-reward='77'", build.toString());
     }
@@ -834,7 +834,7 @@ class BlastNConfigTest
     void test13() {
       var target = new BlastNConfig();
 
-      target.setMaxHSPs((short) 42);
+      target.setMaxHSPs(42);
       assertEquals("-max_hsps='42'", target.toString());
     }
 
@@ -852,7 +852,7 @@ class BlastNConfigTest
     void test15() {
       var target = new BlastNConfig();
 
-      target.setNumDescriptions((short) 42);
+      target.setNumDescriptions(42);
       assertEquals("-num_descriptions='42'", target.toString());
     }
 
@@ -960,7 +960,7 @@ class BlastNConfigTest
     void test27() {
       var target = new BlastNConfig();
 
-      target.setWindowSize((short) 450);
+      target.setWindowSize(450);
       assertEquals("-window_size='450'", target.toString());
     }
 
@@ -1099,7 +1099,7 @@ class BlastNConfigTest
     void test41() {
       var target = new BlastNConfig();
 
-      target.setWordSize((short) 420);
+      target.setWordSize(420);
       assertEquals("-word_size='420'", target.toString());
     }
 
@@ -1144,7 +1144,7 @@ class BlastNConfigTest
     void test46() {
       var target = new BlastNConfig();
 
-      target.setGapOpen((byte) 33);
+      target.setGapOpen(33);
       assertEquals("-gapopen='33'", target.toString());
     }
 
@@ -1153,7 +1153,7 @@ class BlastNConfigTest
     void test47() {
       var target = new BlastNConfig();
 
-      target.setGapExtend((byte) 77);
+      target.setGapExtend(77);
       assertEquals("-gapextend='77'", target.toString());
     }
 
@@ -1162,7 +1162,7 @@ class BlastNConfigTest
     void test48() {
       var target = new BlastNConfig();
 
-      target.setPenalty((byte) 77);
+      target.setPenalty(77);
       assertEquals("-penalty='77'", target.toString());
     }
 
@@ -1171,7 +1171,7 @@ class BlastNConfigTest
     void test49() {
       var target = new BlastNConfig();
 
-      target.setReward((byte) 77);
+      target.setReward(77);
       assertEquals("-reward='77'", target.toString());
     }
